arabic-smart-scribe-main/backend/requirements.txt

fastapi==0.104.1
uvicorn[standard]==0.24.0
google-generativeai==0.5.4
httpx==0.25.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
sqlalchemy==2.0.23
aiofiles==23.2.1

# Database migrations
alembic==1.13.1

# Code quality tools
black==24.4.2
flake8==7.0.0
isort==5.13.2
pytest==7.4.3
pytest-asyncio==0.21.1
pre-commit==3.6.0

# Background tasks
celery[redis]==5.3.4
redis==5.0.1
